[1] merged and removed stages 4 and 6 into 3; [2] reordered remaining stages 5-->3, 3-->4, 7-->5; [3] converted runtimes.md to run_times.rst; [4] removed parks_canada from source variable choices.
