+++ docs/run_times.rst
@@ -1,0 +1,57 @@
+*********
+Run Times
+*********
+
+.. contents::
+   :depth: 2
+
+Prince Edward Island (PE)
+=========================
+
+April 22, 2020
+--------------
+
+:Source: data/raw/pe/PEI_NRN19
+:Environment:
+    :OS: Windows 10 64-bit
+    :CPU: Intel Core i7 4700HQ @ 2.40GHz
+    :RAM:  16.0GB
+    :Python: 3.6.8
+:Runtimes:
+
+=========  ====
+Stage      Time
+=========  ====
+1          08:22.02
+2          02:14.77
+3          02:04.78
+4          01:41.74
+5          02:36.67
+6          00:08.02
+7          11:24.87
+---------  ----
+**Total**  **28:32.87**
+=========  ====
+
+April 27, 2020
+--------------
+
+:Source: data/raw/pe/PEI_NRN19
+:Environment:
+    :OS: Windows 10 64-bit
+    :CPU: Intel Core i7 4700HQ @ 2.40GHz
+    :RAM:  16.0GB
+    :Python: 3.6.8
+:Runtimes:
+
+=========  ====
+Stage      Time
+=========  ====
+1          08:27.93
+2          00:35.56
+3          02:16.30
+4          04:05.12
+5          12:25.98
+---------  ----
+**Total**  **27:50.89**
+=========  ====
