--- docs/runtimes.md
@@ -1,33 +1,0 @@
-# Runtimes
-
-## Prince Edward Island
-
-Date: April 22, 2020  
-Source: data/raw/pe/PEI_NRN19  
-Runtimes:
-
-| Stage | Time |
-| --- | --- |
-1 | 08:22.02
-2 | 02:14.77
-3 | 02:04.78
-4 | 01:41.74
-5 | 02:36.67
-6 | 00:08.02
-7 | 11:24.87
-**Total** | **28:32.87**
-
-Date: April 24, 2020  
-Source: data/raw/pe/PEI_NRN19  
-Runtimes:
-
-| Stage | Time |
-| --- | --- |
-1 | 08:36.01
-2 | 00:41.97
-3 | 02:09.64
-4 | 01:58.95
-5 | 02:05.78
-6 | 00:06.22
-7 | 11:52.25
-**Total** | **27:30.82**
