+++ src/stage_5/distribution_formats/fr/gml.yaml
