+++ src/stage_5/distribution_formats/fr/gpkg.yaml
